+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - feat(init): Initialized new express app with mongodb connection.
 - feat(db): Created mongoose models for the diecast collections.
+- feat(diecast-brands): Created new api to insert diecast brands.
 
 ### Fixed